backend/requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
openai>=1.12.0
ibm-watsonx-ai>=1.1.0
python-dotenv>=1.0.0
gunicorn>=21.2.0
